+++ src/thistle/blake2b.mojo
@@ -7,8 +7,8 @@ RFC 7693
 By Libalpm64, Attribute not required.
 """
 
-from collections import List
-from memory import UnsafePointer, alloc, memcpy
+from std.collections import List
+from std.memory import UnsafePointer, alloc, memcpy
 
 comptime BLAKE2B_IV = SIMD[DType.uint64, 8](
     0x6A09E667F3BCC908,
@@ -39,7 +39,7 @@ comptime SIGMA = (
 
 @always_inline
 fn rotr64[n: Int](x: UInt64) -> UInt64:
-    return (x >> n) | (x << (64 - n))
+    return (x >> UInt64(n)) | (x << UInt64(64 - n))
 
 
 @always_inline
@@ -240,7 +240,7 @@ struct Blake2b(Movable):
             var word_idx = i // 8
             var byte_idx = i % 8
             var word = self.h[word_idx]
-            output.append(UInt8((word >> (byte_idx * 8)) & 0xFF))
+            output.append(UInt8((word >> UInt64(byte_idx * 8)) & 0xFF))
 
         return output^
